@@ -6,16 +6,16 @@
         "@emotion/react": "^11.11.0",
         "@emotion/styled": "^11.11.0",
         "@mui/material": "^5.13.0",
-        "@testing-library/jest-dom": "^5.16.1",
-        "@testing-library/react": "^12.0.0",
+        "@testing-library/jest-dom": "^6.0.0",
+        "@testing-library/react": "^15.0.0",
         "@testing-library/user-event": "^14.4.3",
         "date-fns": "^2.29.1",
         "firebase": "^9.11.0",
         "firebase-tools": "^11.15.0",
         "npm": "^9.0.0",
         "prop-types": "^15.7.2",
-        "react": "^17.0.0",
-        "react-dom": "^17.0.0",
+        "react": "^18.0.0",
+        "react-dom": "^18.0.0",
         "react-router-dom": "^5.2.0",
         "react-scripts": "5.0.1",
         "uuid": "^9.0.0"
