@@ -3,10 +3,13 @@ import './index.css';
 import * as serviceWorker from './utils/serviceWorker';
 
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import Application from './components/Application';
 
-ReactDOM.render(<Application />, document.getElementById('root'));
+const container = document.getElementById('root');
+const root = createRoot(container);
+
+root.render(<Application />);
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
